Send pageleave events using posthog built-in logic
The previous logic was incorrect as it would send the pageleave event on every route change, but not when the user exits the site completely. The pageleave event is intended to be used as a way of knowing when a user has e.g. closed the browser or navigated to a different website.